# Improve performance of NOT IN on indexed properties