--- a/src/realm/query.cpp
+++ b/src/realm/query.cpp
@@ -1695,7 +1695,7 @@ void Query::init() const
 {
     m_table.check();
     if (ParentNode* root = root_node()) {
-        root->init();
+        root->init(m_view == nullptr);
         std::vector<ParentNode*> vec;
         root->gather_children(vec);
     }