The relevant benchmarks:

```
Req runs:   96  QueryChainedOrStrings (MemOnly, EncryptionOff):                   min   5.05ms (-0.43%)            max      6ms (+8.36%)            med   5.17ms (-0.46%)            avg   5.17ms (-0.77%)            stddev   103us (+9.04%)       
Req runs:  231  QueryChainedOrStringsIndexed (MemOnly, EncryptionOff):            min   2.05ms (-0.01%)            max   3.30ms (+29.51%)           med   2.17ms (+1.65%)            avg   2.25ms (+3.88%)            stddev   212us (+91.60%)      
Req runs:   26  QueryNotChainedOrStrings (MemOnly, EncryptionOff):                min  17.80ms (-1.37%)            max  18.97ms (-0.73%)            med  18.25ms (-1.32%)            avg  18.30ms (-1.21%)            stddev   254us (-10.64%)      
Req runs:   26  QueryNotChainedOrStringsIndexed (MemOnly, EncryptionOff):         min  17.95ms (-94.99%)           max  25.28ms (-93.03%)           med  18.47ms (-94.85%)           avg  18.78ms (-94.78%)           stddev  1.39ms (-32.00%)      
Req runs:  144  QueryChainedOrInts (MemOnly, EncryptionOff):                      min   3.37ms (-12.84%)           max   5.07ms (-6.86%)            med   3.44ms (-12.88%)           avg   3.46ms (-13.00%)           stddev   149us (-22.00%)      
Req runs:   15  QueryChainedOrIntsIndexed (MemOnly, EncryptionOff):               min  30.48ms (+0.23%)            max  31.68ms (-1.15%)            med  30.97ms (+0.51%)            avg  31.06ms (+0.77%)            stddev   347us (-6.22%)       
```

We normally don't combine equals nodes inside OR when the expression is on an indexed property because that would change the query from O(M log N) to O(N), which is typically not an improvement. However, inside of a Not group or when a query is limited to a view we evaluate a single row at a time rather than searching ranges, which means that such queries are inherently O(N\*M) regardless of what the query nodes do internally. As a result, we do want to combine OR chains in that case even if there's an index.

Fixes https://github.com/realm/realm-cocoa/issues/4564.